+++ packages/xchain-ethereum/src/index.ts
@@ -1,2 +1,3 @@
 export * from './types'
 export * from './client'
+export { getDefaultFees } from './utils'